+++ buildsystem/dependencies.gradle
@@ -11,7 +11,7 @@ ext {
   serialization_version = '1.2.1'
   coroutines_version = '1.5.0-native-mt'
   sql_delight_version = '1.5.0'
-  klock_version = "2.2.0"
+  kotlinx_datetime_version = "0.2.1"
   stately = "1.1.4"
 }
