+++ common/src/commonMain/kotlin/com.harmony.kotlin/common/date/TimeUnits.kt
@@ -1,0 +1,10 @@
+package com.harmony.kotlin.common.date
+
+/**
+ * Long typealias to provide meaning for second values
+ */
+typealias Seconds = Long
+/**
+ * Long typealias to provide meaning for millisecond values
+ */
+typealias Millis = Long
